+++ packages/BUILD.bazel
@@ -101,6 +101,7 @@ filegroup(
     "//packages/core/saved-objects/core-saved-objects-api-server:build",
     "//packages/core/saved-objects/core-saved-objects-common:build",
     "//packages/core/saved-objects/core-saved-objects-server:build",
+    "//packages/core/test-helpers/core-test-helpers-deprecations-getters:build",
     "//packages/core/test-helpers/core-test-helpers-http-setup-browser:build",
     "//packages/core/theme/core-theme-browser-internal:build",
     "//packages/core/theme/core-theme-browser-mocks:build",
@@ -353,6 +354,7 @@ filegroup(
     "//packages/core/saved-objects/core-saved-objects-api-server:build_types",
     "//packages/core/saved-objects/core-saved-objects-common:build_types",
     "//packages/core/saved-objects/core-saved-objects-server:build_types",
+    "//packages/core/test-helpers/core-test-helpers-deprecations-getters:build_types",
     "//packages/core/test-helpers/core-test-helpers-http-setup-browser:build_types",
     "//packages/core/theme/core-theme-browser-internal:build_types",
     "//packages/core/theme/core-theme-browser-mocks:build_types",
