+++ packages/core/elasticsearch/core-elasticsearch-server-internal/BUILD.bazel
@@ -44,6 +44,7 @@ RUNTIME_DEPS = [
   "//packages/core/http/core-http-server-mocks",
   "//packages/core/elasticsearch/core-elasticsearch-client-server-mocks",
   "//packages/core/config/core-config-server-internal",
+  "//packages/core/test-helpers/core-test-helpers-deprecations-getters",
 ]
 
 TYPES_DEPS = [
@@ -67,6 +68,7 @@ TYPES_DEPS = [
   "//packages/core/execution-context/core-execution-context-server-internal:npm_module_types",
   "//packages/core/elasticsearch/core-elasticsearch-server:npm_module_types",
   "//packages/core/elasticsearch/core-elasticsearch-client-server-internal:npm_module_types",
+  "//packages/core/test-helpers/core-test-helpers-deprecations-getters:npm_module_types"
 ]
 
 jsts_transpiler(
