+++ packages/core/test-helpers/core-test-helpers-deprecations-getters/BUILD.bazel
@@ -2,8 +2,8 @@ load("@npm//@bazel/typescript:index.bzl", "ts_config")
 load("@build_bazel_rules_nodejs//:index.bzl", "js_library")
 load("//src/dev/bazel:index.bzl", "jsts_transpiler", "pkg_npm", "pkg_npm_types", "ts_project")
 
-PKG_DIRNAME = "core-config-server-internal"
-PKG_REQUIRE_NAME = "@kbn/core-config-server-internal"
+PKG_DIRNAME = "core-test-helpers-deprecations-getters"
+PKG_REQUIRE_NAME = "@kbn/core-test-helpers-deprecations-getters"
 
 SOURCE_FILES = glob(
   [
@@ -11,6 +11,7 @@ SOURCE_FILES = glob(
   ],
   exclude = [
     "**/*.test.*",
+    "**/*.stories.*",
   ],
 )
 
@@ -28,9 +29,7 @@ NPM_MODULE_EXTRA_FILES = [
 RUNTIME_DEPS = [
   "//packages/elastic-safer-lodash-set",
   "//packages/kbn-config",
-  "//packages/core/base/core-base-server-internal",
   "//packages/kbn-config-mocks",
-  "//packages/core/test-helpers/core-test-helpers-deprecations-getters",
 ]
 
 TYPES_DEPS = [
@@ -39,8 +38,6 @@ TYPES_DEPS = [
   "//packages/elastic-safer-lodash-set:npm_module_types",
   "//packages/kbn-config:npm_module_types",
   "//packages/kbn-config-mocks:npm_module_types",
-  "//packages/core/base/core-base-server-internal:npm_module_types",
-  "//packages/core/test-helpers/core-test-helpers-deprecations-getters:npm_module_types",
 ]
 
 jsts_transpiler(
