+++ packages/core/test-helpers/core-test-helpers-deprecations-getters/README.md
@@ -1,0 +1,3 @@
+# @kbn/core-test-helpers-deprecations-getters
+
+Empty package generated by @kbn/generate
