+++ packages/core/test-helpers/core-test-helpers-deprecations-getters/src/deprecations_getters.ts
